refactor(core): 迁移 query 管理路由到 admin 域
