+++ noj-core/src/domains/admin/routes/query.ts
@@ -1,6 +1,6 @@
 import { Hono } from "hono";
 import type { AuthEnv } from "./../../identity/index.ts";
-import { getDashboardStats } from "../services/dashboard.ts";
+import { getDashboardStats } from "../../query/services/dashboard.ts";
 import { getObservabilitySnapshot } from "../../system/services/observability.ts";
 
 /**
